% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exelbib.R
\name{add_xlsx}
\alias{add_xlsx}
\title{Add an Empty Excel-File and Start using Excel as Reference Software}
\usage{
add_xlsx(xlsx_name = "References.xlsx")
}
\arguments{
\item{xlsx_name}{The name you want to use for your Excel-file where you organize your references.}
}
\description{
Add an Empty Excel-File and Start using Excel as Reference Software
}
\examples{
add_xlsx(xlsx_name= "Excel_References_example")
}
