% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exelbib.R
\name{bib_to_xlsx}
\alias{bib_to_xlsx}
\title{Title}
\usage{
bib_to_xlsx(
  bib_file_path = "export.bib",
  excel_file_path = "Excel_References.xlsx",
  sheet_name = "Import"
)
}
\arguments{
\item{bib_file_path}{The name of your .bib-file. Defaults to "export.bib".}

\item{excel_file_path}{The name or directory of the Excel file you use for your References. In the file, the bibtex-entries should be arranged in a column, one entry per cell.}

\item{sheet_name}{Sheet name where you want the references imported to, default is "Import".}
}
\value{
References in your Excel file!
}
\description{
Title
}
\examples{
download.file(url="https://raw.githubusercontent.com/Enno-W/excelbib/main/Excel_References.xlsx", destfile = "Excel_References.xlsx", mode ="wb")
download.file(url="https://raw.githubusercontent.com/Enno-W/excelbib/main/export.bib", destfile = "export.bib", mode ="wb")
require("openxlsx")
bib_to_xlsx()
}
