% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exelbib.R
\name{process_bib_to_excel}
\alias{process_bib_to_excel}
\title{Title}
\usage{
process_bib_to_excel(
  bib_file_path = "export.bib",
  excel_file_path = "Excel_References.xlsx",
  sheet_name = "Import"
)
}
\arguments{
\item{sheet_name}{}
}
\value{
References in your Excel file!
}
\description{
Title
}
\examples{
download.file(url="https://raw.githubusercontent.com/Enno-W/excelbib/main/Excel_References.xlsx", destfile = "Excel_References.xlsx", mode ="wb")
download.file(url="https://raw.githubusercontent.com/Enno-W/excelbib/main/export.bib", destfile = "Excel_References.xlsx", mode ="wb")
}
