% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exelbib.R
\name{xlsx_to_bib}
\alias{xlsx_to_bib}
\title{Update your References from an Excel-file}
\usage{
xlsx_to_bib(
  excel_file = "References.xlsx",
  bib_file = "bibliography.bib",
  sheet = 2,
  column = 1,
  first_row = 2
)
}
\arguments{
\item{excel_file}{The name or directory of the Excel file you use for your References. In the file, the bibtex-entries should be arranged in a column, one entry per cell.}

\item{bib_file}{The name of your .bib-file. Defaults to "bibliography.bib". The file will be added to your working directory.}

\item{sheet}{The sheet number from which you want to import your references. Defaults to 2.}

\item{column}{The column from which you want to import your references. The default is 1.}

\item{first_row}{The first row that contains a bibtex entry that you want to cite.}
}
\value{
An error message if the file does not exist.
}
\description{
Update your References from an Excel-file
}
\examples{
download.file(url="https://raw.githubusercontent.com/Enno-W/excelbib/main/Excel_References_example.xlsx", destfile = "Excel_References_example.xlsx", mode ="wb")
require("openxlsx")
xlsx_to_bib ("Excel_References_example.xlsx", bib_file="example_bibliography.bib", sheet = 2, column = 1, first_row = 2)
}
